Change the behavior of org-typst-preview
Previously, org-typst-preview toggle the preview of the Typst code
fragment which is pointed by the cursor, this will trigger an error
when the cursor is not inside a Typst code fragment. Now,
org-typst-preview toggle the preview of the Typst code fragment that
is closest to the cursor, if there is any. Otherwise, it emits an
message in the echo area.